Update solution - CodeHub

+++ 7-kyu/most-digits/most-digits.java
@@ -1,13 +1,13 @@
  
 public class MostDigits {
   public static int findLongest(int[] numbers) {
-    int result =0;
+    long result =0;
     for(long i=0; i<numbers.length; i++){
       if(result<numbers[i]){
         result = numbers[i];
       }
     }
-      return result;
+      return (int)result;
   }
 ​
 }
